+++ src/structures/vroom/job.h
@@ -34,7 +34,8 @@ struct Job {
   Job(Id id,
       const Location& location,
       Duration service = 0,
-      const Amount& amount = Amount(0),
+      const Amount& delivery = Amount(0),
+      const Amount& pickup = Amount(0),
       const Skills& skills = Skills(),
       Priority priority = 0,
       const std::vector<TimeWindow>& tws =