+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- `pickup` and `delivery` keys at `job` level in input (#262)
 - `priority` key for jobs to gain some control on which jobs are unassigned (#246)
 - `HttpWrapper` class to factor code previously duplicated across routing wrappers (#224)
 
@@ -12,6 +13,10 @@
 - Speed up solving by 25% for CVRP and up to 30% for VRPTW benchmark instances (#255)
 - Update Travis configuration to use Ubuntu Bionic (#260)
 - Cut down validity checks time (#266)
+
+### Deprecated
+
+- `amount` key at `job` level in input and at `summary` and `route` level in output (#262)
 
 ### Removed